update package.json and del create-directory
