+++ package.json
@@ -29,7 +29,6 @@
     "@react-three/fiber": "^6.1.2",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
-    "strip-ansi": "^3.0.1",
     "three": "^0.128.0"
   }
 }
