+++ packages/create-video/package.json
@@ -26,7 +26,6 @@
 		"@jonny/eslint-config": "^2.1.233",
 		"@types/fs-extra": "^9.0.11",
 		"@types/node": "^15.0.1",
-		"@types/npm-package-arg": "^6.1.0",
 		"@types/prompts": "^2.0.12",
 		"eslint": "^7.15.0",
 		"ts-node": "^9.1.1",
